Move collections controller override to a monkey patch
